Cambios en la migración.

+++ database/migrations/2018_12_09_155032_alter_facturas_ventas_cab.php
@@ -25,8 +25,11 @@ class AlterFacturasVentasCab extends Migration
      */
     public function down()
     {
-        Schema::table('facturas_ventas_cab', function (Blueprint $table) {
-            $table->dropColumn('nume_serie');
-        });
+        //Pablo - 14/10/2018 - controlamos si existen las columnas primero antes del drop        
+        if(Schema::hasColumn('facturas_ventas_cab','nume_serie') ) {
+            Schema::table('facturas_ventas_cab', function (Blueprint $table) {
+                $table->dropColumn('nume_serie');
+            });
+        }
     }
 }
